added support for keyboard shortcuts, and cleanup
